+++ tsconfig.json
@@ -18,10 +18,7 @@
     "verbatimModuleSyntax": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./*"],
-      "@/hooks/*": ["lib/hooks/*"],
-      "@/types/*": ["lib/types/*"],
-      "@/utils/*": ["lib/utils/*"]
+      "@/*": ["./*"]
     },
     "plugins": [
       {
